+++ handlers/databaseHandler.js
@@ -1,0 +1,13 @@
+const mongoose = require("mongoose");
+
+function config() {
+    if (process.env.MONGO_CONNECTION_URL) {
+        mongoose.connect(process.env.MONGO_CONNECTION_URL, {
+            useNewUrlParser: true,
+            useUnifiedTopology: true,
+            useFindAndModify: false,
+        });
+    }
+}
+
+module.exports.config = config;